+++ R/misc_utils.R
@@ -1673,20 +1673,6 @@ insertMetaInfo <- function(str){
 
 
 
-CheckDetailsTablePerformed <-function(type){
-  paramSet <- readSet(paramSet, "paramSet");
-  dataSet <- readDataset(paramSet$dataName);
-
-  performed <- T;
-  if(grepl("bmd", type)){
-    performed <- !is.null(dataSet$html.resTable);
-  }
-
-  #print(paste("checkPerformed=", type, "====",performed));
-
-return(performed)
-}
-
 BuildCEMiNet <- function(dataName,
                          filter      = TRUE,
                          min_ngen    = 30,